Support reuse and do not reuse connection for bolt. (#509)
* support reuse and do not reuse the connection for bolt
